Replaces release workflow with ncipollo/release-action.
The actions/create-release action is unmaintained.
It seems strange that creating a tag and creating a release are two different operations, requiring the tag creator to have first updated the changelog.
The new action also creates the tag, meaning we only have to update the changelog and manually run release.